move undoFormatIndexName to ITElasticSearchClien
